pyproject.toml
[tool.poetry]
name = "dspy-multi-lingual-optimizers"
version = "0.1.0"
description = ""
authors = ["Adrien <adrien.bertheleme@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dspy-ai = "^2.4.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py311"


[tool.ruff.isort]
case-sensitive = true

[tool.ruff.lint]
select = [
  # bugbear rules
  "B",
  "I",
  # remove unused imports
  "F401",
  # bare except statements
  "E722",
  # unused arguments
  "ARG",
]
ignore = [
  "B006",
  "B018",
]

unfixable = [
  "T201",
  "T203",
]
ignore-init-module-imports = true
